Move the 'AddNodePool' components into their own folder
